Fix 64bit dlog format warnings

+++ src/util/mtp_fs.c
@@ -779,7 +779,7 @@ mtp_bool _util_get_filesystem_info_ext(mtp_char *storepath,
 	}
 
 	capacity = used_size = avail_size = (mtp_uint64)buf.f_bsize;
-	DBG("Block size : %d\n", buf.f_bsize);
+	DBG("Block size : %lu\n", (unsigned long)buf.f_bsize);
 	capacity *= buf.f_blocks;
 	used_size *= (buf.f_blocks - buf.f_bavail);
 	avail_size *= buf.f_bavail;
